Move legacy Glyph extensions to the Visual Studio layer
Thankfully the only things still using the old glyph types (versus
the new image monikers) are already in the Visual Studio layer, so
this is an easy move.
